Extract ToolRuntimeContext into tool_surface/runtime_context.rs
